Solution to Code Academy while loops
Goal: Ask user to guess a number in the range 1-9.  The user has three
guesses to do so.

Code Used:
from: usually precedes the import command
import: a command used to incorporate a module, which is like a
pre-written program, into the current program
randrange( , ): generates a random number from the left to the right
while : one of two words used to signify a loop
break: ends a loop